Rebased and cleaned up repository (marker)
This commit contains no changes, as it's simply a recordkeeping marker
to represent that I just went through and cleaned out the history of
this repository, removing a few large files and committed copies of life
files, as well as fixing the committers to be the latest emails. I tried
to leave as much of my humanity, late-night commits and mistakes as I
could while still removing the data-leak level things. :)

This makes this repository ready to publish to github. Enjoy!